\name{mktdata_skeleton-package}
\alias{mktdata_skeleton-package}
\alias{mktdata_skeleton}
\docType{package}
\title{
\packageTitle{mktdata_skeleton}
}
\description{
\packageDescription{mktdata_skeleton}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mktdata_skeleton}
\packageIndices{mktdata_skeleton}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mktdata_skeleton}

Maintainer: \packageMaintainer{mktdata_skeleton}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
